Handle pre-existing linearize destination folder
Fixes a bug which cause a problem if a master repo was linearized
into an already existing folder containing a linearized version
of that repo.
In this new version, this situation is detected and the linearization
is aborted.
A new command line option is provided to automatically delete an
already existing destination folder