Clean-up printing of error messages
Until now, an error resulted in a fast exit of the running program.

With the introduction of a master repo validation command, we want to
print out all found error and hence, a quick exit is not desirable.

Instead a new method 'printError' has been introduced that takes
an implicit parameter that specifies whether exit on error applies
or not