Make cpCleanViaGit more robust
Using a static name for the temporary remote on the master git repo is
not very robust. In case the cleanup is not performed (for example
in case the execution of 'studentify' or 'linearize' is interrupted
by the user), subsequent runs of 'studentify' or 'linearize' will
fail systematically. With this change, the added remote gets a random
name (UUID) prefixed by the string 'CMT-'.