Fix issue with exercises in root folder studentified repo
When CMT configuration setting `studentified-repo-active-exercise-folder` is
set to `"."`, the exercise code is put in the root folder of the studentified
artefact when studentifying a course.

This commit fixes the problem that overwrote the CMT configuration folder
in the studentified artifact. The fix consists of tuning the `cmt-studentified-dont-touch`
configuration setting in the studentified artifact.

For example, with this main CMT configuration file:

```
  studentified-repo-active-exercise-folder = code

  cmt-studentified-dont-touch = [
    .idea
    .bsp
    .bloop
    .vscode
    .metals
    project/target
    project/project
  ]
```

the resulting studentified artifact configuration settings are:

```
    "active-exercise-folder" : "code",
    "cmt-studentified-dont-touch" : [
        "code/.idea",
        "code/.bsp",
        "code/.bloop",
        "code/.vscode",
        "code/.metals",
        "code/project/target",
        "code/project/project"
    ],
```

In contrast, with the following config file:

```
  studentified-repo-active-exercise-folder = .

  cmt-studentified-dont-touch = [
    .idea
    .bsp
    .bloop
    .vscode
    .metals
    project/target
    project/project
  ],
```

the resulting studentified artifact configuration settings look
as follows:

```
    "active-exercise-folder" : ".",
    "cmt-studentified-dont-touch" : [
        ".idea",
        ".bsp",
        ".bloop",
        ".vscode",
        ".metals",
        "project/target",
        "project/project",
        ".cmt",
        ".git",
        ".gitignore"
    ],
```